[#158142427] Move tokens to a different module
