+++ src/controllers/__tests__/messagesController.test.ts
@@ -4,11 +4,11 @@ import { ResponseSuccessJson } from "italia-ts-commons/lib/responses";
 import mockReq from "../../__mocks__/request";
 import mockRes from "../../__mocks__/response";
 import ApiClient from "../../services/apiClientFactory";
-import { SessionToken, WalletToken } from "../../services/ISessionStorage";
 import MessagesService from "../../services/messagesService";
 import { EmailAddress } from "../../types/api/EmailAddress";
 import { FiscalCode } from "../../types/api/FiscalCode";
 import { SpidLevelEnum } from "../../types/api/SpidLevel";
+import { SessionToken, WalletToken } from "../../types/token";
 import { User } from "../../types/user";
 import MessagesController from "../messagesController";
