+++ src/types/token.ts
@@ -1,0 +1,14 @@
+import * as t from "io-ts";
+import { tag } from "italia-ts-commons/lib/types";
+
+interface ISessionTokenTag {
+  readonly kind: "SessionToken";
+}
+export const SessionToken = tag<ISessionTokenTag>()(t.string);
+export type SessionToken = t.TypeOf<typeof SessionToken>;
+
+interface IWalletTokenTag {
+  readonly kind: "WalletToken";
+}
+export const WalletToken = tag<IWalletTokenTag>()(t.string);
+export type WalletToken = t.TypeOf<typeof WalletToken>;
